requirements.txt
Django==4
static3==0.7.0
gunicorn==19.6.0
django-extensions==3
dj-database-url==2
djangorestframework>=3.14
mysqlclient
geocoder
pycountry